==> 01_data_extraction.ipynb <==
## Import packages

import pandas as pd
import os

## Import OSHA data
The imported data contains:
* Case Id 
* Title / summary of accident
* Description of accident
* Keywords/ tags of important words to be highlighted
* Victims' information

osha_filename = os.path.join(os.getcwd(), 'input_data', 'osha.xlsx')
df = pd.read_excel(osha_filename, names=['case_id', 'title', 'description', 'keywords', 'victims'], header=None)
df = df.set_index('case_id', drop=True)
df.head()

df.describe()

## Remove invalid data
* For those records with description = 'InspectionOpen DateSICEstablishment Name'
* For those records with title is numeric

incomplete1 = df['description'].str.lower().str.contains('InspectionOpen DateSICEstablishment Name'.lower())
incomplete2 = df['title'].apply(lambda x: str(x).isnumeric())
df = df[~incomplete1 & ~incomplete2]
df.head()

df.describe()

## Export data to csv

export_file = os.path.join(os.getcwd(), 'output_data', '01_data_extraction_final.csv')
df.to_csv(export_file)
