==> 05_analysis.ipynb <==
## Import packages

import os
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer

## Import data

file_input = os.path.join(os.getcwd() + os.sep, 'output_data', '03_topics_modeling_final.csv')
df = pd.read_csv(file_input, encoding='utf-8')
df.head()

# flattern "occupation, body_parts"
def create_dtm(data):
    
    df = data.copy()
    df['occupation'] = df['occupation'].apply(lambda x: "occupation_" + str(x))
    df['occupation'] = df['occupation'].apply(lambda x: str(x).replace(", ", " occupation_"))
    df['body_parts'] = df['body_parts'].apply(lambda x: "body_" + str(x))
    df['body_parts'] = df['body_parts'].apply(lambda x: str(x).replace(", ", " body_"))
    df['occupation_body'] = df['occupation'].map(str) + " " + df['body_parts'].map(str)
    
    count_vect = CountVectorizer( )
    X_counts = count_vect.fit_transform(df.occupation_body)
    dtm = pd.DataFrame(X_counts.toarray().transpose(), index = count_vect.get_feature_names())
    #dtm = dtm.transpose()
    
    dtm_ = pd.concat([df, dtm.transpose()], axis=1)
    del dtm_['occupation']
    del dtm_['body_parts']
    del dtm_['occupation_body']
    
    return dtm_

df_final = create_dtm(df)
df_final.to_csv("05_analysis_forBI.csv",sep=",")
df_final.head()
